Further reduce code duplication and move utils classes to services
