+++ src/controllers/certificates/delivery.details.controller.ts
@@ -9,7 +9,7 @@ import { createLogger } from "ch-structured-logging";
 import { APPLICATION_NAME } from "../../config/config";
 import { deliveryDetailsValidationRules, validate } from "../../utils/delivery-details-validation";
 import { setServiceUrl } from "../../utils/service.url.utils";
-import { renderNonBasketJourneyDeliveryDetails, setBackLink } from "../../utils/delivery.details.utils";
+import { renderNonBasketJourneyDeliveryDetails, setBackLink } from "../../service/delivery.details.service";
 import { DeliveryDetailsPropertyName } from "./model/DeliveryDetailsPropertyName";
 const escape = require("escape-html");
